Refactor to visually separate client from common code
