+++ src/main/java/org/github/melodiccougar7/immersivefixes/mixin/common/CokeOvenStateAccessor.java
@@ -1,4 +1,4 @@
-package org.github.melodiccougar7.immersivefixes.mixin;
+package org.github.melodiccougar7.immersivefixes.mixin.common;
 
 import blusunrize.immersiveengineering.common.blocks.multiblocks.logic.CokeOvenLogic;
 import org.spongepowered.asm.mixin.Mixin;
