+++ src/main/java/org/github/melodiccougar7/immersivefixes/mixin/common/DieselGenFixes.java
@@ -1,4 +1,4 @@
-package org.github.melodiccougar7.immersivefixes.mixin;
+package org.github.melodiccougar7.immersivefixes.mixin.common;
 
 import blusunrize.immersiveengineering.api.energy.GeneratorFuel;
 import blusunrize.immersiveengineering.api.multiblocks.blocks.component.IClientTickableComponent;
