+++ src/main/resources/immersivefixes.mixins.json
@@ -6,20 +6,20 @@
   "compatibilityLevel": "JAVA_17",
   "refmap": "immersivefixes.refmap.json",
   "mixins": [
-    "ArcFurnaceSecondariesChanceMixin",
-    "CokeOvenLogicFix",
-    "CokeOvenStateAccessor",
-    "DieselGenAccessor",
-    "DieselGenFixes",
-    "MixinFluidProtection"
+    "common.ArcFurnaceSecondariesChanceMixin",
+    "common.CokeOvenLogicFix",
+    "common.CokeOvenStateAccessor",
+    "common.DieselGenAccessor",
+    "common.DieselGenFixes",
+    "common.MixinFluidProtection"
   ],
   "client": [
-    "gui.CircuitTableSceenMixin",
-    "gui.GuiButtonLogicCircuitRegisterMixins",
-    "gui.ScreenInvoker",
-    "ClientEventHandlerMixin",
-    "IEMuffledSoundMixin",
-    "SoundEngineMixin"
+    "client.gui.CircuitTableSceenMixin",
+    "client.gui.GuiButtonLogicCircuitRegisterMixins",
+    "client.gui.ScreenInvoker",
+    "client.ClientEventHandlerMixin",
+    "client.IEMuffledSoundMixin",
+    "client.SoundEngineMixin"
 
   ],
   "injectors": {
